backend/requirements.txt
fastapi~=0.82.0
pandas~=1.4.3
requests~=2.26.0
python-multipart==0.0.5
uvicorn~=0.18.3
pyyaml~=6.0
scikit-learn~=1.1.1
lightgbm~=3.3.2
optuna~=2.10.1
joblib~=1.1.0
numpy~=1.20.3
